refactor(weave): move otel_traces_endpoint to weave.trace.urls
URL/path construction belongs alongside the existing
remote_project_root_url / redirect_call helpers in
weave/trace/urls.py, not in weave_init.py. Moves
_OTEL_GENAI_TRACES_PATH and otel_traces_endpoint() and updates
the two import sites (weave/__init__.py public re-export,
weave_init._setup_session_tracing internal call).

Pure refactor; no behavior change.
