+++ weave/__init__.py
@@ -94,9 +94,9 @@ from weave.session.session import (
     start_turn,
 )
 from weave.trace.log_call import log_call
+from weave.trace.urls import otel_traces_endpoint
 from weave.trace.util import Thread as Thread
 from weave.trace.util import ThreadPoolExecutor as ThreadPoolExecutor
-from weave.trace.weave_init import otel_traces_endpoint
 from weave.type_handlers.Audio.audio import Audio
 from weave.type_handlers.File.file import File
 from weave.type_handlers.Markdown.markdown import Markdown
