+++ weave/trace/urls.py
@@ -6,6 +6,26 @@ from weave.trace import env
 BROWSE3_PATH = "browse3"
 WEAVE_SLUG = "weave"
 
+# OTel GenAI traces ingestion path on the weave trace server. Appended to
+# ``weave_trace_server_url()`` to form the OTLP HTTP exporter endpoint.
+_OTEL_GENAI_TRACES_PATH = "/agents/otel/v1/traces"
+
+
+def otel_traces_endpoint(base_url: str | None = None) -> str:
+    """Return the full OTLP HTTP endpoint URL for Weave GenAI trace ingestion.
+
+    External callers (e.g. boot-time probes that want to verify the
+    ingest endpoint is reachable before relying on the BatchSpanProcessor
+    to silently drop exports) should call this rather than constructing
+    the URL by hand. The path is owned by the SDK and may move.
+
+    Args:
+        base_url: Trace server base URL. Defaults to
+            ``weave_trace_server_url()``.
+    """
+    server_url = (base_url or env.weave_trace_server_url()).rstrip("/")
+    return f"{server_url}{_OTEL_GENAI_TRACES_PATH}"
+
 
 def remote_project_root_url(entity_name: str, project_name: str) -> str:
     return f"{wandb.app_url(env.wandb_base_url())}/{entity_name}/{quote(project_name)}"
